### 2018-11-07 #1068 @arfon
@katyhuff - would you be willing to edit this submission for JOSS?
