### 2018-11-07 #1068 @katyhuff
@mikaem, your suggestions would be great.

Meanwhile -- perhaps @lucydot or @lindsayad would be interested. @lucydot, @lindsayad, would either of you be willing to be reviewers for this JOSS submission?
