+++ docs/paper/paper.bib
@@ -45,13 +45,13 @@ eprint = {https://doi.org/10.1137/0915089}
   eprint    = {1701.03787}
 }
 
-@article{mortensen17,
+@inproceedings{mortensen17,
 	author = {Mortensen, Mikael},
-	booktitle = {MekIT - Ninth national conference on Computational Mechanics},
+	booktitle = {MekIT'17 - Ninth national conference on Computational Mechanics},
 	isbn = {978-84-947311-1-2},
 	pages = {273--298},
 	publisher = {International Center for Numerical Methods in Engineering (CIMNE)},
 	title = {Shenfun - automating the spectral Galerkin method},
-	ty = {CHAPTER},
+  editor = {Skallerud, Bjorn Helge and Andersson, Helge Ingolf},
 	year = {2017}
   }
