### 2018-11-08 #1068 @lindsayad
I can review if another reviewer is still needed

> On Nov 8, 2018, at 3:14 AM, Lucy Whalley <notifications@github.com> wrote:
>
> @katyhuff @mikaem I’d be very happy to review.
>
