### 2018-11-08 #1071 @whedon
[REVIEW]: Shenfun: High Performance Computing Platform for the Spectral Galerkin method
**Submitting author:** @mikaem (<a href="http://orcid.org/0000-0002-3293-7573">Mikael Mortensen</a>)
**Repository:** <a href="https://github.com/spectralDNS/shenfun" target ="_blank">https://github.com/spectralDNS/shenfun</a>
**Version:** 1.2.1
