### 2018-11-09 1854309
Make all write/read close/open files every time. Fixing in demos
