+++ demo/neumann_poisson1D.py
@@ -14,7 +14,8 @@ whereas for Chebyshev we solve
      (\nabla^2 u, v) = (f, v)
 
 """
-import sys, os
+import sys
+import os
 import importlib
 from sympy import symbols, sin, lambdify
 import numpy as np
