+++ tests/conftest.py
@@ -1,5 +1,5 @@
-import pytest
 import os
+import pytest
 
 def pytest_configure(config):
     os.environ['pytest'] = 'True'
