### 2018-11-09 1b2ca9a
Make all write/read close/open files every time. Fixing in demos
