+++ demo/neumann_poisson2D.py
@@ -16,17 +16,14 @@ whereas for Chebyshev we solve
      (\nabla^2 u, v) = (f, v)
 
 """
-import sys, os
+import sys
+import os
 import importlib
 from sympy import symbols, cos, sin, lambdify
 import numpy as np
+from mpi4py import MPI
 from shenfun import inner, div, grad, TestFunction, TrialFunction, \
     TensorProductSpace, Basis, Array, Function
-from mpi4py import MPI
-try:
-    import matplotlib.pyplot as plt
-except ImportError:
-    plt = None
 
 comm = MPI.COMM_WORLD
 
@@ -37,7 +34,7 @@ Solver = base.la.Helmholtz
 
 # Use sympy to compute a rhs, given an analytical solution
 x, y = symbols("x,y")
-ue =  cos(4*y)*sin(2*np.pi*x)*(1-x**2)
+ue = cos(4*y)*sin(2*np.pi*x)*(1-x**2)
 fe = ue.diff(x, 2) + ue.diff(y, 2)
 
 # Lambdify for faster evaluation
@@ -81,7 +78,8 @@ uj = ul(*X)
 print(abs(uj-uq).max())
 assert np.allclose(uj, uq)
 
-if plt is not None and not 'pytest' in os.environ:
+if 'pytest' not in os.environ:
+    import matplotlib.pyplot as plt
     plt.figure()
     plt.contourf(X[0], X[1], uq)
     plt.colorbar()
@@ -96,4 +94,3 @@ if plt is not None and not 'pytest' in os.environ:
     plt.title('Error')
 
     plt.show()
-
