+++ demo/fourier_Helmholtz3D.py
@@ -18,13 +18,9 @@ where V is the Fourier basis :math:`span{exp(1jkx)}_{k=-N/2}^{N/2-1}` and
 import os
 import numpy as np
 from sympy import symbols, cos, sin, lambdify
+from mpi4py import MPI
 from shenfun import inner, div, grad, TestFunction, TrialFunction, Basis, \
     TensorProductSpace, Array, Function
-from mpi4py import MPI
-try:
-    import matplotlib.pyplot as plt
-except ImportError:
-    plt = None
 
 comm = MPI.COMM_WORLD
 
@@ -64,9 +60,10 @@ uj = ul(*X)
 print(abs(uj-uq).max())
 assert np.allclose(uj, uq)
 
-if plt is not None and not 'pytest' in os.environ:
+if 'pytest' not in os.environ:
+    import matplotlib.pyplot as plt
     plt.figure()
-    plt.contourf(X[0][:, :, 0], X[1][: ,: ,0], uq[:, :, 0])
+    plt.contourf(X[0][:, :, 0], X[1][:, :, 0], uq[:, :, 0])
     plt.colorbar()
 
     plt.figure()
@@ -78,43 +75,3 @@ if plt is not None and not 'pytest' in os.environ:
     plt.colorbar()
     plt.title('Error')
     #plt.show()
-
-#from shenfun import VectorTensorProductSpace, curl, project
-
-#Tk = VectorTensorProductSpace(T)
-#v = TestFunction(Tk)
-##u_ = Function(Tk, False)
-##u_[:] = np.random.random(u_.shape)
-##u_hat = Function(Tk)
-##u_hat = Tk.forward(u_, u_hat)
-##w_hat = inner(v, curl(u_), uh_hat=u_hat)
-
-##u0 = u_[0]
-##inner(v, u_)
-
-#uq = T.as_function(uq)
-#du_hat = Function(Tk)
-#f_hat = T.as_function(f_hat)
-#du_hat = project(grad(uq), Tk, output_array=du_hat, uh_hat=f_hat)
-#du = Function(Tk, False)
-#du = Tk.backward(du_hat, du)
-
-#dux = ue.diff(x, 1)
-#duxl = lambdify((x, y, z), dux, 'numpy')
-#duxj = duxl(*X)
-#duy = ue.diff(y, 1)
-#duyl = lambdify((x, y, z), duy, 'numpy')
-#duyj = duyl(*X)
-#duz = ue.diff(z, 1)
-#duzl = lambdify((x, y, z), duz, 'numpy')
-#duzj = duzl(*X)
-
-
-#plt.figure()
-#plt.contourf(X[0][:,:,0], X[1][:,:,0], du[0, :, :, 0])
-#plt.colorbar()
-
-##plt.show()
-#assert np.allclose(duxj, du[0])
-#assert np.allclose(duyj, du[1])
-#assert np.allclose(duzj, du[2])
