### 2018-11-09 b879609
Make all shapes and slices return tuples an not lists
