### 2018-11-09 87eefd4
Make all shapes and slices return tuples an not lists
