### 2018-11-11 #1071 @lucydot
In general, shenfun is packaged and documented very nicely. I need a few more days to look at functionality, but here are my comments so far:

## Documentation

- I really like the paper published on RawGit  (Unfortunately I noticed that RawGit is shutting down next year, so you'll have to move it across to Github Pages or similar at some point).
- It might be useful to state in the readme that shenfun is a Python 3 package (compatible with Python 2?)
- I think you could move a very short summary of Shenfun's unique selling points (global shape functions which lead to more accurate approximations than FEniCS / ability to run DNS on supercomputers using an accessible high-level language) to the readme and/or landing page of read the docs.
- You could give instructions for running Pytest locally `python -m pytest` and explicitly point to where the Travis builds are.
- You could include a "how to cite" section in your documentation (Zenodo DOI until the JOSS paper is published), including a bibtex entry.
- There are not clear guidelines for people who might want to contribute to the software.

## Installation

- I used conda to install mpi4py and FFTW, followed by pip install shenfun. There was a problem with pip as cython was not installed - I think this is because it is under  `setup_requires` rather than `install_requires` in the `setup.py` file, so pip ignores it:  https://github.com/pytest-dev/pytest-xdist/issues/136. After pip installing cython, pip install shenfun worked fine.

## Functionality

- I've been working through the Klein-Gordon equation demo. I've ran on up to 24 cores, doubling the cores gives ~1.5X speed-up, so the MPI looks like it is working good. I'd like a few more days to play with the functionality before ticking it off.
