### 2018-11-11 #1071 @lindsayad
I've been very impressed by the documentation, and I've enjoyed my first exposure to a spectral element implementation.

I'm also exploring the Klein-Gordon problem and I'm a little curious about some of the timings. I'm running on my laptop which has four real cpus, capable of hyperthreading. Results:
```
(spectralDNS) lindad@localhost:~/scratch$ time mpiexec -np 1 python ./klein-gordon.py

real	0m15.007s
user	0m14.711s
sys	0m0.273s
(spectralDNS) lindad@localhost:~/scratch$ time mpiexec -np 2 python ./klein-gordon.py

real	0m11.127s
user	0m19.624s
sys	0m0.520s
(spectralDNS) lindad@localhost:~/scratch$ time mpiexec -np 4 python ./klein-gordon.py

real	0m9.876s
user	0m33.977s
sys	0m1.127s
(spectralDNS) lindad@localhost:~/scratch$ time mpiexec -np 8 python ./klein-gordon.py

real	0m11.531s
user	1m12.890s
sys	0m5.291s
```
So I do see a speed-up while increasing my real cpu count, but a stagnation or performance decline when using hyperthreading. The latter doesn't necessarily mean too much to me for reasons like [this](https://superuser.com/a/279803). However, I guess I'm a little curious about the relatively small speed-up from 1 to 4 cores. Is there a lot of serial computation or communication? I'm also curious how @lucydot did her timings? I also understand that scaling studies are fraught with peril, especially for novice users of software.
