### 2018-11-11 #1071 @lindsayad
Sorry, much better to use the `time` module as it's already used in the solver script. Limiting output to rank 0:
```
(spectralDNS) lindad@localhost:~/scratch$ mpiexec -np 8 python ./klein-gordon.py
Time  8.530541181564331
(spectralDNS) lindad@localhost:~/scratch$ mpiexec -np 4 python ./klein-gordon.py
Time  7.4511120319366455
(spectralDNS) lindad@localhost:~/scratch$ mpiexec -np 2 python ./klein-gordon.py
Time  8.857075214385986
(spectralDNS) lindad@localhost:~/scratch$ mpiexec -np 1 python ./klein-gordon.py
Time  14.115585088729858
```
So very strong performance gain moving from 1 to 2 procs, but seemingly diminishing returns moving from 2 to 4. With four procs, there should still be roughly `64**3 / 4 = 65536` degrees of freedom per process, so it seems like still plenty of work for all. Any comment on this? It's not something I need to dwell on, but curious.
