### 2018-11-12 #1071 @mikaem
@lucydot Thanks a lot for very good and constructive feedback! I'll start by answering some of your questions/comments below:

Documentation
---------------

    - I really like the paper published on RawGit (Unfortunately I noticed that RawGit is shutting down next year, so you'll have to move it across to Github Pages or similar at some point).

Thanks for the heads up, I did not know that. I'll look into other options.

    - It might be useful to state in the readme that shenfun is a Python 3 package (compatible with Python 2?)

Yes, I think you're right. A short section with dependencies could be in the readme. I'll add it.

    - I think you could move a very short summary of Shenfun's unique selling points (global shape functions which lead to more accurate approximations than FEniCS / ability to run DNS on supercomputers using an accessible high-level language) to the readme and/or landing page of read the docs.

Very good idea:-)

    - You could give instructions for running Pytest locally python -m pytest and explicitly point to where the Travis builds are.
    - You could include a "how to cite" section in your documentation (Zenodo DOI until the JOSS paper is published), including a bibtex entry.
    - There are not clear guidelines for people who might want to contribute to the software.

All good and valid points. I'll add these at appropriate locations to the documentation.
