### 2018-11-12 #1071 @mikaem
@lindsayad Thank you for the fast and nice feedback:-)
Some comments on performance. Spectral methods make use of global basis functions and the communication load is therefore very high (MPI Alltoall), much larger than for example for a finite element method that only needs to communicate data on the interface between distributed meshes. For this reason it is usually necessary with a good computer (high performance supercomputer preferably:-)) with fast interconnect between CPUs to see good speedup. Most laptops do not have good enough interconnect speeds to achieve good scaling, but a [Cray XC](https://www.cray.com/products/computing/xc-series) does:-).

For a laptop with 4 cores (like my own 3 year old MacBook pro) you should be able to get speedup all the way up to 4, though, even if not perfect. However, you should use keyword `slab=True` when creating the TensorProductSpace

    T = TensorProductSpace(comm, (K0, K1, K2), slab=True)

This has to do with how the data are decomposed and distributed between processors. A slab uses one processor group whereas with `slab=False` you will use as many groups as possible, here 2. The advantage with using 2 groups is that you can use many more processors than only for one group. But if you are only going to use 4 processors, then one group is actually faster. That may explain your results that are using 4 processors. Another reason can be the hardware itself. Because most processors with 4 cores usually come with two sockets, with two cores in each socket. And communication inside the same socket is faster than between sockets. And as such you may get a nice speedup going from one to two, but not as good moving to four. With a supercomputer the communication speeds are much higher than in your laptop and speedup should be more or less perfect up to thousands of processors. In the end this all depends on the hardware.

Hyperthreading is a different story. You need to activate it using, e.g.,

    T = TensorProductSpace(comm, (K0, K1, K2), slab=True, **{'threads': 2})

This will use the FFTW library with hyperthreading (OpenMP). Again, speedup is very hardware dependent; I have seen speedup on some computers, but nothing on others. I'm not an expert though.
