### 2018-11-12 #1071 @mikaem
@lucydot Regarding installation, what was the problem with Cython? My understanding is that if cython is in `setup_requires`, then pip will install it if it is missing? And cython is only required for building the code, not for running shenfun, which is why I did not include it under `install_requires`.
