### 2018-11-12 #1071 @mikaem
@lucydot Regarding functionality please see my response to @lindsayad. For scaling tests please use `slab=True`
