### 2018-11-12 e7e2ab8
Update README.rst
