### 2018-11-12 ce38377
Update README.rst
