### 2018-11-12 1436cec
Update README.rst
