### 2018-11-12 e1181de
Working on documentation after comments in JOSS review
