### 2018-11-12 5a55a9a
Working on documentation after comments in JOSS review
