+++ docs/source/index.rst
@@ -6,12 +6,22 @@
 Welcome to shenfun's documentation!
 ===================================
 
-`Shenfun`_ is a high performance computing platform for solving partial
-differential equations by the spectral Galerkin method.
+Shenfun is a high performance computing platform for solving partial
+differential equations (PDEs) by the spectral Galerkin method.
 The user interface to shenfun is very similar to `FEniCS`_,
 but applications are limited to multidimensional tensor product grids.
-The code is parallelized with MPI through the `mpi4py-fft`_
-package.
+The code is parallelized with MPI through the `mpi4py-fft`_ package.
+
+Shenfun enables you to develop efficient and accurate PDE solvers
+(spectral order and accuracy), in the comfortable high-level Python
+language. The spectral accuracy is ensured from using high-order
+*global* orthogonal basis functions (Fourier, Legendre and Chebyshev),
+as opposed to finite element codes like `FEniCS`_ that are using low-order
+*local* basis functions. Efficiency is ensured through vectorization
+(`Numpy`_), parallelization (`mpi4py`_) and by moving critical routines to
+`Cython`_. Shenfun has been used to run turbulence simulations (Direct
+Numerical Simulations) on thousands of processors on high-performance
+supercomputers, see the `spectralDNS`_ repository.
 
 Document build status
 ---------------------
@@ -27,6 +37,8 @@ Document build status
    introduction
    gettingstarted
    installation
+   howtocite
+   howtocontribute
 
 .. toctree::
    :caption: Demos:
@@ -45,4 +57,8 @@ Document build status
 
 .. _shenfun: https:/github.com/spectralDNS/shenfun
 .. _mpi4py-fft: https://bitbucket.org/mpi4py/mpi4py-fft
+.. _mpi4py: https://bitbucket.org/mpi4py/mpi4py
 .. _Fenics: https://fenicsproject.org
+.. _Numpy: https:/www.numpy.org
+.. _Cython: https://cython.org
+.. _spectralDNS: https://github.com/spectralDNS/spectralDNS
