+++ docs/source/howtocontribute.rst
@@ -1,0 +1,17 @@
+How to contribute?
+==================
+
+Shenfun is an open source project and anyone is welcome to contribute.
+An easy way to get started is by suggesting a new enhancement on the
+`issue tracker <https://github.com/spectralDNS/shenfun/issues>`_. If you have
+found a bug, then either report this on the issue tracker, er even better, make
+a fork of the repository, fix the bug and then create a
+`pull request <https://github.com/spectralDNS/shenfun/pulls>`_
+to get the fix into the master branch.
+
+If you have a particularily interesting application, then we are very interested
+in pull requests that are adding new demo programs. If you feel like adding
+an extended demo, then have a look at the examples in the
+`docs/demos folder <https://github.com/spectralDNS/shenfun/tree/master/docs/demos>`_.
+Note that extended demos are written using
+`doconce <http://hplgit.github.io/doconce/doc/web/index.html>`_.
