+++ docs/source/howtocite.rst
@@ -1,0 +1,19 @@
+How to cite?
+============
+
+Please cite shenfun using
+
+::
+
+    @inproceedings{shenfun,
+    	author = {Mortensen, Mikael},
+    	booktitle = {MekIT'17 - Ninth national conference on Computational Mechanics},
+    	isbn = {978-84-947311-1-2},
+    	pages = {273--298},
+        archivePrefix = "arXiv",
+        eprint = {1708.03188},
+    	publisher = {International Center for Numerical Methods in Engineering (CIMNE)},
+    	title = {Shenfun - automating the spectral Galerkin method},
+    	editor = {Skallerud, Bjorn Helge and Andersson, Helge Ingolf},
+    	year = {2017}
+    }
