### 2018-11-12 #1071 @lindsayad
Thanks for the comprehensive reply :-) I’m satisfied with it.

Regarding the installation, I ran into the same issue as @lucydot. Cython was not installed automatically by shenfun when pip installing the latter. After manually installing cython first (as noted in your documentation) shenfun obviously installed fine.

> On Nov 12, 2018, at 3:45 AM, Mikael Mortensen <notifications@github.com> wrote:
>
> @lucydot Regarding functionality please see my response to @lindsayad. For scaling tests please use slab=True
>
