### 2018-11-12 #1071 @mikaem
Ok. Should be fixed by adding cython to `install_requires`, right?
