### 2018-11-12 #1071 @lucydot
Hi @mikaem, thank you for your post re: scaling up. I was running the calculations on HPC which explains why I got a better scale-up. Yes, moving to `install_requires` should fix it.
