### 2018-11-13 29adf47
cython -> install_requires + modified CachedArrayDict to allow for not filling the array
