### 2018-11-13 13d60d4
cython -> install_requires + modified CachedArrayDict to allow for not filling the array
