### 2018-11-14 1925f66
Fixing references in documentation (doconce setting)
