+++ docs/demos/KleinGordon/Makefile
@@ -13,20 +13,19 @@ ${name}.pdf: ${name}.tex
 	pdflatex -shell-escape ${name}
 
 ${name}.html: ${name}.do.txt
-	doconce format html ${name} --html_style=${style} --pygments_html_style=default --html_admon=bootstrap_alert --html_output=${name}_${style} --keep_pygments_html_bg —html_code_style=inherit --html_pre_style=inherit 
+	doconce format html ${name} --html_style=${style} --pygments_html_style=default --html_admon=bootstrap_alert --html_output=${name}_${style} --keep_pygments_html_bg —html_code_style=inherit --html_pre_style=inherit
 	doconce split_html ${name}_${style}.html
 
 ${name}.rst: ${name}.do.txt
-	doconce format sphinx ${name}
+	doconce format sphinx ${name} --sphinx_preserve_bib_keys
 
 ${name}.ipynb: ${name}.do.txt
 	doconce format ipynb ${name}
-  
-view: 
+
+view:
 	open ${name}.pdf &
 
 ##########################################################################3
 
 clean:
 	latexmk -C
-
