### 2018-11-14 764bf56
Fixing references in documentation (doconce setting)
