+++ docs/demos/Poisson/poisson.do.txt
@@ -1,14 +1,16 @@
-TITLE: Demo - 1D Poisson equation 
+TITLE: Demo - 1D Poisson equation
 AUTHOR: Mikael Mortensen {copyright|CC BY} Email:mikaem@math.uio.no at Department of Mathematics, University of Oslo.
 DATE: today
 
 __Summary.__
 This is a demonstration of how the Python module "shenfun":
 "https://github.com/spectralDNS/shenfun" can be used to solve the Poisson
-equation with Dirichlet boundary conditions in one dimension. Spectral convergence, as shown in Figure ref{fig:ct0}, is demonstrated. 
+equation with Dirichlet boundary conditions in one dimension. Spectral convergence, as
+shown in Figure ref{fig:ct0}, is demonstrated.
 The demo is implemented in
 a single Python file "dirichlet_poisson1D.py":
-"https://github.com/spectralDNS/shenfun/blob/master/demo/dirichlet_poisson1D.py", and the numerical method is is described in more detail by J. Shen cite{shen1} and cite{shen95}.
+"https://github.com/spectralDNS/shenfun/blob/master/demo/dirichlet_poisson1D.py", and
+the numerical method is is described in more detail by J. Shen cite{shen1} and cite{shen95}.
 
 FIGURE: [https://rawgit.com/spectralDNS/spectralutilities/master/figures/poisson1D_errornorm.png] Convergence of 1D Poisson solvers for both Legendre and Chebyshev modified basis function. label{fig:ct0}
 
@@ -44,17 +46,17 @@ def func(x):
 
 The Poisson equation is given as
 
-!bt 
+!bt
 \begin{align}
 \nabla^2 u(x) &= f(x) \quad \text{for }\, x \in [-1, 1], label{eq:poisson}\\
 u(-1)&=a, u(1)=b, \notag
 \end{align}
 !et
 where $u(x)$ is the solution, $f(x)$ is a function and $a, b$ are two possibly
-non-zero constants. 
+non-zero constants.
 
-To solve Eq. (ref{eq:poisson}) with the Galerkin method we need smooth continuously 
-differentiable basis functions, $v_k$, that satisfy the given boundary conditions. 
+To solve Eq. (ref{eq:poisson}) with the Galerkin method we need smooth continuously
+differentiable basis functions, $v_k$, that satisfy the given boundary conditions.
 And then we look for solutions like
 !bt
 \begin{equation}
@@ -62,12 +64,12 @@ u(x) = \sum_{k=0}^{N} \hat{u}_k v_k(x), label{eq:u}
 \end{equation}
 !et
 
-where $N+1$ is the size of the discretized problem and the basis is 
+where $N+1$ is the size of the discretized problem and the basis is
 $V^N=\text{span}\{v_k\}_{k=0}^{N}$.
-The basis functions can, for example,  be constructed from 
-"Chebyshev": "https://en.wikipedia.org/wiki/Chebyshev_polynomials", $T_k(x)$, or 
-"Legendre": "https://en.wikipedia.org/wiki/Legendre_polynomials", $L_k(x)$, functions 
-and we use the common notation $\phi_k(x)$ to represent either one of them. It turns out that 
+The basis functions can, for example,  be constructed from
+"Chebyshev": "https://en.wikipedia.org/wiki/Chebyshev_polynomials", $T_k(x)$, or
+"Legendre": "https://en.wikipedia.org/wiki/Legendre_polynomials", $L_k(x)$, functions
+and we use the common notation $\phi_k(x)$ to represent either one of them. It turns out that
 it is easiest to use basis functions with homogeneous Dirichlet boundary conditions
 !bt
 \begin{equation}
@@ -93,21 +95,21 @@ label{eq:dirichletb}
 !et
[…]
