+++ docs/source/poisson.rst
@@ -7,12 +7,14 @@ Demo - 1D Poisson equation
 %%%%%%%%%%%%%%%%%%%%%%%%%%
 
 :Authors: Mikael Mortensen (mikaem at math.uio.no)
-:Date: Nov 12, 2018
+:Date: Nov 14, 2018
 
 *Summary.* This is a demonstration of how the Python module `shenfun <https://github.com/spectralDNS/shenfun>`__ can be used to solve the Poisson
-equation with Dirichlet boundary conditions in one dimension. Spectral convergence, as shown in Figure :ref:`fig:ct0`, is demonstrated. 
+equation with Dirichlet boundary conditions in one dimension. Spectral convergence, as
+shown in Figure :ref:`fig:ct0`, is demonstrated.
 The demo is implemented in
-a single Python file `dirichlet_poisson1D.py <https://github.com/spectralDNS/shenfun/blob/master/demo/dirichlet_poisson1D.py>`__, and the numerical method is is described in more detail by J. Shen [Ref1]_ and [Ref2]_.
+a single Python file `dirichlet_poisson1D.py <https://github.com/spectralDNS/shenfun/blob/master/demo/dirichlet_poisson1D.py>`__, and
+the numerical method is is described in more detail by J. Shen :cite:`shen1` and :cite:`shen95`.
 
 .. _fig:ct0:
 
@@ -43,10 +45,10 @@ The Poisson equation is given as
         
 
 where :math:`u(x)` is the solution, :math:`f(x)` is a function and :math:`a, b` are two possibly
-non-zero constants. 
+non-zero constants.
 
-To solve Eq. :eq:`eq:poisson` with the Galerkin method we need smooth continuously 
-differentiable basis functions, :math:`v_k`, that satisfy the given boundary conditions. 
+To solve Eq. :eq:`eq:poisson` with the Galerkin method we need smooth continuously
+differentiable basis functions, :math:`v_k`, that satisfy the given boundary conditions.
 And then we look for solutions like
 
 .. math::
@@ -56,12 +58,12 @@ And then we look for solutions like
         u(x) = \sum_{k=0}^{N} \hat{u}_k v_k(x), 
         
 
-where :math:`N+1` is the size of the discretized problem and the basis is 
+where :math:`N+1` is the size of the discretized problem and the basis is
 :math:`V^N=\text{span}\{v_k\}_{k=0}^{N}`.
-The basis functions can, for example,  be constructed from 
-`Chebyshev <https://en.wikipedia.org/wiki/Chebyshev_polynomials>`__, :math:`T_k(x)`, or 
-`Legendre <https://en.wikipedia.org/wiki/Legendre_polynomials>`__, :math:`L_k(x)`, functions 
-and we use the common notation :math:`\phi_k(x)` to represent either one of them. It turns out that 
+The basis functions can, for example,  be constructed from
+`Chebyshev <https://en.wikipedia.org/wiki/Chebyshev_polynomials>`__, :math:`T_k(x)`, or
+`Legendre <https://en.wikipedia.org/wiki/Legendre_polynomials>`__, :math:`L_k(x)`, functions
+and we use the common notation :math:`\phi_k(x)` to represent either one of them. It turns out that
 it is easiest to use basis functions with homogeneous Dirichlet boundary conditions
 
 .. math::
@@ -111,14 +113,14 @@ of freedom, :math:`\hat{u}_{N-1}` and :math:`\hat{u}_{N}`, now are given as
 
 and, as such, we only have to solve for :math:`\{\hat{u}_k\}_{k=0}^{N-2}`, just like
[…]
