### 2018-11-16 8e22a17
Microoptimizations - avoiding tests and function calls on slice dict lookup
