+++ shenfun/chebyshev/bases.py
@@ -6,7 +6,8 @@ import numpy as np
 from numpy.polynomial import chebyshev as n_cheb
 from scipy.special import eval_chebyt
 from mpi4py_fft import fftw
-from shenfun.spectralbase import SpectralBase, work, Transform, FuncWrap
+from shenfun.spectralbase import SpectralBase, work, Transform, FuncWrap, \
+    islicedict, slicedict
 from shenfun.optimization import Cheb
 from shenfun.utilities import inheritdocstrings
 
@@ -194,6 +195,8 @@ class ChebyshevBase(SpectralBase):
         self.forward = Transform(self.forward, xfftn_fwd, U, V, V)
         self.backward = Transform(self.backward, xfftn_bck, V, V, U)
         self.scalar_product = Transform(self.scalar_product, xfftn_fwd, U, V, V)
+        self.si = islicedict(axis=self.axis, dimensions=self.dimensions())
+        self.sl = slicedict(axis=self.axis, dimensions=self.dimensions())
 
 
 @inheritdocstrings
@@ -266,11 +269,10 @@ class Basis(ChebyshevBase):
         return fd.copy()
 
     def apply_inverse_mass(self, array):
-        sl = self.sl
         array *= (2/np.pi)
-        array[sl(0)] /= 2
+        array[self.si[0]] /= 2
         if self.quad == 'GL':
-            array[sl(-1)] /= 2
+            array[self.si[-1]] /= 2
         return array
 
     def evaluate_expansion_all(self, input_array, output_array, fast_transform=True):
@@ -279,18 +281,18 @@ class Basis(ChebyshevBase):
             return
 
         output_array = self.backward.xfftn()
-        sl = self.sl
         if self.quad == "GC":
+            s0 = self.sl[slice(0, 1)]
             output_array *= 0.5
-            output_array += input_array[sl(slice(0, 1))]/2
+            output_array += input_array[s0]/2
 
         elif self.quad == "GL":
             output_array *= 0.5
-            output_array += input_array[sl(slice(0, 1))]/2
-            s0 = sl(slice(-1, None))
-            s2 = sl(slice(0, None, 2))
+            output_array += input_array[self.sl[slice(0, 1)]]/2
+            s0 = self.sl[slice(-1, None)]
+            s2 = self.sl[slice(0, None, 2)]
             output_array[s2] += input_array[s0]/2
-            s2 = sl(slice(1, None, 2))
+            s2 = self.sl[slice(1, None, 2)]
             output_array[s2] -= input_array[s0]/2
 
     def evaluate_scalar_product(self, input_array, output_array, fast_transform=True):
@@ -378,26 +380,23 @@ class ShenDirichletBasis(ChebyshevBase):
             self.vandermonde_scalar_product(input_array, output_array)
             return
         output = self.CT.scalar_product(fast_transform=fast_transform)
-        s0 = self.sl(0)
-        s1 = self.sl(1)
+        s0 = self.si[0]
+        s1 = self.si[1]
         c0 = 0.5*(output[s0] + output[s1])
         c1 = 0.5*(output[s0] - output[s1])
-        s0 = self.sl(slice(0, -2))
-        s1 = self.sl(slice(2, None))
+        s0 = self.sl[slice(0, -2)]
+        s1 = self.sl[slice(2, None)]
         output[s0] -= output[s1]
-        s0 = list(s0)
-        s0[self.axis] = -2
-        output[tuple(s0)] = c0
-        s0[self.axis] = -1
-        output[tuple(s0)] = c1
+        output[self.si[-2]] = c0
+        output[self.si[-1]] = c1
 
[…]
