### 2018-11-16 4fc77be
Microoptimizations - avoiding tests and function calls on slice dict lookup
