+++ shenfun/legendre/bases.py
@@ -7,7 +7,8 @@ import numpy as np
 from numpy.polynomial import legendre as leg
 from scipy.special import eval_legendre
 from mpi4py_fft import fftw
-from shenfun.spectralbase import SpectralBase, work, Transform
+from shenfun.spectralbase import SpectralBase, work, Transform, islicedict, \
+    slicedict
 from shenfun.utilities import inheritdocstrings
 from .lobatto import legendre_lobatto_nodes_and_weights
 
@@ -124,6 +125,8 @@ class LegendreBase(SpectralBase):
         self.forward = Transform(self.forward, None, U, V, V)
         self.backward = Transform(self.backward, None, V, V, U)
         self.scalar_product = Transform(self.scalar_product, None, U, V, V)
+        self.si = islicedict(axis=self.axis, dimensions=self.dimensions())
+        self.sl = slicedict(axis=self.axis, dimensions=self.dimensions())
 
 
 @inheritdocstrings
@@ -216,8 +219,8 @@ class ShenDirichletBasis(LegendreBase):
     #                           fast_transform=False): # pragma: no cover
     #    # Not used since there are no fast transforms for Legendre
     #    w_hat = work[(input_array, 0)]
-    #    s0 = self.sl(slice(0, -2))
-    #    s1 = self.sl(slice(2, None))
+    #    s0 = self.sl[slice(0, -2)]
+    #    s1 = self.sl[slice(2, None)]
     #    self.set_factor_array(input_array)
     #    w_hat[s0] = input_array[s0]*self._factor
     #    w_hat[s1] -= input_array[s0]*self._factor
@@ -264,6 +267,8 @@ class ShenDirichletBasis(LegendreBase):
         self.forward = Transform(self.forward, None, U, V, V)
         self.backward = Transform(self.backward, None, V, V, U)
         self.scalar_product = Transform(self.scalar_product, None, U, V, V)
+        self.si = islicedict(axis=self.axis, dimensions=self.dimensions())
+        self.sl = slicedict(axis=self.axis, dimensions=self.dimensions())
 
 
 @inheritdocstrings
@@ -312,8 +317,8 @@ class ShenNeumannBasis(LegendreBase):
     def scalar_product(self, input_array=None, output_array=None, fast_transform=False):
         output = SpectralBase.scalar_product(self, input_array, output_array, False)
 
-        output[self.sl(0)] = self.mean*np.pi
-        output[self.sl(slice(-2, None))] = 0
+        output[self.si[0]] = self.mean*np.pi
+        output[self.sl[slice(-2, None)]] = 0
         return output
 
     def evaluate_basis(self, x, i=0, output_array=None):
@@ -327,8 +332,8 @@ class ShenNeumannBasis(LegendreBase):
     #    # Not used since there are no fast transforms for Legendre
     #    w_hat = work[(input_array, 0)]
     #    self.set_factor_array(input_array)
-    #    s0 = self.sl(slice(0, -2))
-    #    s1 = self.sl(slice(2, None))
+    #    s0 = self.sl[slice(0, -2)]
+    #    s1 = self.sl[slice(2, None)]
     #    w_hat[s0] = input_array[s0]
     #    w_hat[s1] -= self._factor*input_array[s0]
     #    output_array = self.LT.backward(w_hat)
@@ -363,7 +368,8 @@ class ShenNeumannBasis(LegendreBase):
         self.forward = Transform(self.forward, None, U, V, V)
         self.backward = Transform(self.backward, None, V, V, U)
         self.scalar_product = Transform(self.scalar_product, None, U, V, V)
-
[…]
