### 2018-11-16 367fcce
Fixing dimensions for collapse_fourier
