### 2018-11-19 #1071 @lindsayad
Amazing that the system "std" C++ headers for OS X Sierra are not standard compliant (real and imaginary component setters are done through references instead of through void methods that take an argument (as the standard dictates)). But after using a local standard library install, I get all tests to pass with the exception of the `test_curl` test for long double, e.g. I get a `KeyError` for `'G'`. I don't understand this failure...my intuition suggests that the error is probably not caused by Shenfun. Anyway with 4727 passes, I'm pretty much ready to sign off on this myself. Tagging @katyhuff
