### 2018-11-19 #1071 @katyhuff
Very nice everyone! @lindsayad @lucydot thanks a lot for your quick and thorough reviews. I'll take a quick look over in the morning and we'll finalize the acceptance.
