### 2018-11-19 #1071 @katyhuff
@mikaem thank you for a strong submission and for engaging actively in the review process!
Could you please make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission! (You may want to be sure to double check the paper, all minor details, etc. before creating the archive).

Thank you again, @lucydot and @lindsayad for the reviews! I know it's a lot of work, and I appreciate that you both conducted very prompt and thorough reviews.
