### 2018-11-19 #1071 @mikaem
Thanks a lot @katyhuff, @lucydot and @lindsayad :-)

Is it too late to make any changes to the pdf? I noticed that changes to the documentation based on the very good suggestions of the esteemed reviewers had not been migrated to the pdf:-)
