### 2018-11-19 #1071 @lindsayad
Thanks, I enjoyed reviewing!

> On Nov 19, 2018, at 6:39 AM, whedon <notifications@github.com> wrote:
>
> 👉 Check article proof 📄 👈
>
