### 2018-11-19 c42a20c
Cleaning up cython code
