+++ tests/test_matrices.py
@@ -30,11 +30,19 @@ formats = ('dia', 'cython', 'python', 'self')
 N = 16
 k = np.arange(N).astype(float)
 a = np.random.random(N)
-b = np.random.random((N, N, N))
 c = np.zeros(N)
 c1 = np.zeros(N)
-d = np.zeros((N, N, N))
-d1 = np.zeros((N, N, N))
+
+work = {
+    3:
+    (np.random.random((N, N, N)),
+     np.zeros((N, N, N)),
+     np.zeros((N, N, N))),
+    2:
+    (np.random.random((N, N)),
+     np.zeros((N, N)),
+     np.zeros((N, N)))
+    }
 
 cbases2 = list(product(cBasis, cBasis))
 lbases2 = list(product(lBasis, lBasis))
@@ -60,9 +68,9 @@ def test_mat(key, mat, quad):
 @pytest.mark.parametrize('b0,b1', cbases2)
 @pytest.mark.parametrize('quad', cquads)
 @pytest.mark.parametrize('format', formats)
-@pytest.mark.parametrize('axis', (0, 1, 2))
+@pytest.mark.parametrize('dim', (2, 3))
 @pytest.mark.parametrize('k', range(5))
-def test_cmatvec(b0, b1, quad, format, axis, k):
+def test_cmatvec(b0, b1, quad, format, dim, k):
     """Test matrix-vector product"""
     global c, c1, d, d1
     b0 = b0(N, quad=quad)
@@ -71,29 +79,30 @@ def test_cmatvec(b0, b1, quad, format, axis, k):
     c = mat.matvec(a, c, format='csr')
     c1 = mat.matvec(a, c1, format=format)
     assert np.allclose(c, c1)
+    for axis in range(0, dim):
+        b, d, d1 = work[dim]
+        d.fill(0)
+        d1.fill(0)
+        d = mat.matvec(b, d, format='csr', axis=axis)
+        d1 = mat.matvec(b, d1, format=format, axis=axis)
+        assert np.allclose(d, d1)
 
-    d.fill(0)
-    d1.fill(0)
-    d = mat.matvec(b, d, format='csr', axis=axis)
-    d1 = mat.matvec(b, d1, format=format, axis=axis)
-    assert np.allclose(d, d1)
-
-    # Test multidimensional with axis equals 1D case
-    d1.fill(0)
-    bc = [np.newaxis,]*3
-    bc[axis] = slice(None)
-    fj = np.broadcast_to(a[tuple(bc)], (N,)*3).copy()
-    d1 = mat.matvec(fj, d1, format=format, axis=axis)
-    cc = [0,]*3
-    cc[axis] = slice(None)
-    assert np.allclose(c, d1[tuple(cc)])
+        # Test multidimensional with axis equals 1D case
+        d1.fill(0)
+        bc = [np.newaxis,]*dim
+        bc[axis] = slice(None)
+        fj = np.broadcast_to(a[tuple(bc)], (N,)*dim).copy()
+        d1 = mat.matvec(fj, d1, format=format, axis=axis)
+        cc = [0,]*dim
+        cc[axis] = slice(None)
+        assert np.allclose(c, d1[tuple(cc)])
 
 @pytest.mark.parametrize('b0,b1', lbases2)
 @pytest.mark.parametrize('quad', lquads)
 @pytest.mark.parametrize('format', formats)
-@pytest.mark.parametrize('axis', (0, 1, 2))
+@pytest.mark.parametrize('dim', (2, 3))
 @pytest.mark.parametrize('k0,k1', product((0, 1, 2), (0, 1, 2)))
[…]
