### 2018-11-19 a4d8ac0
Cleaning up cython code
