+++ shenfun/optimization/la.pyx
@@ -866,45 +866,101 @@ def LU_Helmholtz_1D(A, B,
             d2[i] = A_scale*A_4[i] - L[i-2]*d2[i-2]
 
 
-def Solve_Helmholtz_1D(np.ndarray[T, ndim=1] fk,
-                       np.ndarray[T, ndim=1] u_hat,
-                       bint neumann,
-                       np.ndarray[real_t, ndim=1] d0,
-                       np.ndarray[real_t, ndim=1] d1,
-                       np.ndarray[real_t, ndim=1] d2,
-                       np.ndarray[real_t, ndim=1] L):
+def LU_Helmholtz_3D(A, B, np.int64_t axis,
+                    np.ndarray[real_t, ndim=3] A_scale,
+                    np.ndarray[real_t, ndim=3] B_scale,
+                    bint neumann,
+                    np.ndarray[real_t, ndim=3] d0,
+                    np.ndarray[real_t, ndim=3] d1,
+                    np.ndarray[real_t, ndim=3] d2,
+                    np.ndarray[real_t, ndim=3] L):
     cdef:
-        int i, j, N
+        unsigned int ii, jj
+
+    if axis == 0:
+        for ii in range(d0.shape[1]):
+            for jj in range(d0.shape[2]):
+                LU_Helmholtz_1D(A, B,
+                                A_scale[0, ii, jj],
+                                B_scale[0, ii, jj],
+                                neumann,
+                                d0[:, ii, jj],
+                                d1[:, ii, jj],
+                                d2[:, ii, jj],
+                                L [:, ii, jj])
+
+    elif axis == 1:
+        for ii in range(d0.shape[0]):
+            for jj in range(d0.shape[2]):
+                LU_Helmholtz_1D(A, B,
+                                A_scale[ii, 0, jj],
+                                B_scale[ii, 0, jj],
+                                neumann,
+                                d0[ii, :, jj],
+                                d1[ii, :, jj],
+                                d2[ii, :, jj],
+                                L [ii, :, jj])
+
+    elif axis == 2:
+        for ii in range(d0.shape[0]):
+            for jj in range(d0.shape[1]):
+                LU_Helmholtz_1D(A, B,
+                                A_scale[ii, jj, 0],
+                                B_scale[ii, jj, 0],
+                                neumann,
+                                d0[ii, jj, :],
+                                d1[ii, jj, :],
+                                d2[ii, jj, :],
+                                L [ii, jj, :])
+
+def LU_Helmholtz_2D(A, B, np.int64_t axis,
+                    np.ndarray[real_t, ndim=2] A_scale,
+                    np.ndarray[real_t, ndim=2] B_scale,
+                    bint neumann,
+                    np.ndarray[real_t, ndim=2] d0,
+                    np.ndarray[real_t, ndim=2] d1,
+                    np.ndarray[real_t, ndim=2] d2,
+                    np.ndarray[real_t, ndim=2] L):
+    cdef:
+        unsigned int ii
+
+    if axis == 0:
+        for ii in range(d0.shape[1]):
+            LU_Helmholtz_1D(A, B,
+                            A_scale[0, ii],
+                            B_scale[0, ii],
+                            neumann,
+                            d0[:, ii],
+                            d1[:, ii],
[…]
