+++ shenfun/optimization/Matvec.pyx
@@ -4,7 +4,7 @@ import numpy as np
 cimport cython
 cimport numpy as np
 from libcpp.vector cimport vector
-from libc.math cimport M_PI
+from libc.math cimport M_PI, M_PI_2
 from cython.parallel import prange
 
 ctypedef np.complex128_t complex_t
@@ -25,142 +25,328 @@ def imult(T[:, :, ::1] array, real_t scale):
                 array[i, j, k] *= scale
     return array
 
-
-def CDNmat_matvec(np.ndarray[real_t, ndim=1] ud,
-                  np.ndarray[real_t, ndim=1] ld,
-                  np.ndarray[T, ndim=3] v,
-                  np.ndarray[T, ndim=3] b,
-                  np.int64_t axis):
+def CDN_matvec1D_ptr(T[::1] v,
+                     T[::1] b,
+                     real_t[::1] ld,
+                     real_t[::1] ud):
     cdef:
-        int i, j, k
+        T* v_ptr = &v[0]
+        T* b_ptr = &b[0]
         int N = v.shape[0]-2
+    CDN_matvec_ptr(v_ptr, b_ptr, &ld[0], &ud[0], N, 1)
 
-    if axis == 0:
-        for j in xrange(b.shape[1]):
-            for k in xrange(b.shape[2]):
-                b[0, j, k] = ud[0]*v[1, j, k]
-                b[N-1, j, k] = ld[N-2]*v[N-2, j, k]
-
-        for i in xrange(1, N-1):
-            for j in xrange(b.shape[1]):
-                for k in xrange(b.shape[2]):
-                    b[i, j, k] = ud[i]*v[i+1, j, k] + ld[i-1]*v[i-1, j, k]
-
-    elif axis == 1:
-        for i in xrange(b.shape[0]):
-            for k in xrange(b.shape[2]):
-                b[i, 0, k] = ud[0]*v[i, 1, k]
-                b[i, N-1, k] = ld[N-2]*v[i, N-2, k]
-
-        for i in xrange(b.shape[0]):
-            for j in xrange(1, N-1):
-                for k in xrange(b.shape[2]):
-                    b[i, j, k] = ud[j]*v[i, j+1, k] + ld[j-1]*v[i, j-1, k]
-
-    elif axis == 2:
-        for i in xrange(b.shape[0]):
-            for j in xrange(b.shape[1]):
-                b[i, j, 0] = ud[0]*v[i, j, 1]
-                b[i, j, N-1] = ld[N-2]*v[i, j, N-2]
-
-        for i in xrange(b.shape[0]):
-            for j in xrange(b.shape[1]):
-                for k in xrange(1, N-1):
-                    b[i, j, k] = ud[k]*v[i, j, k+1] + ld[k-1]*v[i, j, k-1]
-
-
-def BDNmat_matvec(real_t ud,
-                  np.ndarray[real_t, ndim=1] ld,
-                  np.ndarray[real_t, ndim=1] dd,
-                  np.ndarray[T, ndim=3] v,
-                  np.ndarray[T, ndim=3] b,
-                  np.int64_t axis):
+cdef void CDN_matvec_ptr(T* v,
+                         T* b,
+                         real_t* ld,
+                         real_t* ud,
+                         int N,
+                         int st):
     cdef:
-        int i, j, k
-        int N = v.shape[0]-2
+        int i
 
-    if axis == 0:
-        for j in xrange(b.shape[1]):
-            for k in xrange(b.shape[2]):
[…]
