### 2018-11-19 #1071 @katyhuff
@mikaem it is not at all too late to make changes to the pdf. Quite the opposite -- the ideal time is now. So, go ahead and double check that paper, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. When you update this thread with the DOI of the archive, I'll move forward with the submission. Until then, this is your moment for final touchups!
