### 2018-11-19 6adb0a9
Update paper.md
