### 2018-11-19 2c34fd9
Update paper.md
