### 2018-11-19 #1071 @mikaem
@katyhuff is it also ok to bump the version number before a final submit? I used version 1.2.0 in the original submission, but there has been modifications.
