### 2018-11-19 #1071 @katyhuff
Indeed @mikaem, if a new release is desired, to reflect review-related changes to the paper and code, many authors do issue a new (minor or patch) release version at this stage, before making the DOI.
