### 2018-11-19 1ea7365
bump version number for joss release
