### 2018-11-19 #1071 @mikaem
Thanks @katyhuff, we should then be able to continue
DOI that resolves to latest version:
https://doi.org/10.5281/zenodo.1237749
DOI of latest version now:
https://doi.org/10.5281/zenodo.1491713
