+++ conf/conda/meta.yaml
@@ -6,7 +6,7 @@ source:
   git_url: ../../
 
 build:
-  number: 0
+  number: 1
 
 requirements:
   build:
